Backend/requirements.txt
django 
djangorestframework
celery 
redis
django-cors-headers
django-filter
pillow
psycopg2-binary
django-redis
djoser
python-decouplepip
"psycopg[binary]"